ostree-commit-push.sh: Synchronize local OSTree repository from server before make a new OSTree commit
Synchronize from the OSTree Repository Server to the local OSTree
repository before make the root filesystem as a new OSTree commit. So,
local OSTree repository has the difference between the latest commit
from the server and the new commit. And, ostree-push can only transfer
the difference to the OSTree repository server.